Factor out engine folder
